+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.2.14-rc.0
 - added support for dispenser redirect
+- fix for custom redirect button on "claiming finished" page
 
 ## 2.2.13
 - added event for "Powered by Linkdrop" click